pyproject.toml
[project]
name = "reviewhood"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django-rest-framework>=0.1.0",
    "django>=5.1.4",
    "python-dotenv>=1.0.1",
    "psycopg2>=2.9.10",
    "djangorestframework-simplejwt>=5.3.1",
]
